refactor: FE-2039 deprecate form, pod and spinner
