+++ change_log/next/FE-2039-deprecate-form-component.yml
@@ -1,0 +1,6 @@
+Deprecated Components: "`Form` has been moved to `__deprecated__/components/form`,
+  to continue using it you will need to update the import path to `carbon-react/lib/__deprecated__/components/form` \n
+  `Pod` has been moved to `__deprecated__/components/pod`,
+  to continue using it you will need to update the import path to `carbon-react/lib/__deprecated__/components/pod` \n
+  `Spinner` has been moved to `__deprecated__/components/spinner`,
+  to continue using it you will need to update the import path to `carbon-react/lib/__deprecated__/components/spinner`"
