+++ demo/definitions.js
@@ -27,7 +27,7 @@ export default {
   'fieldset':             require('__deprecated__/components/fieldset/__definition__').default,
   'filter':               require('components/filter/__definition__').default,
   'flash':                require('components/flash/__definition__').default,
-  'form':                 require('components/form/__definition__').default,
+  'form':                 require('__deprecated__/components/form/__definition__').default,
   'grouped-character':    require('__deprecated__/components/grouped-character/__definition__').default,
   'heading':              require('components/heading/__definition__').default,
   'help':                 require('components/help/__definition__').default,
@@ -45,7 +45,7 @@ export default {
   'pager':                require('components/pager/__definition__').default,
   'pages':                require('components/pages/__definition__').default,
   'pill':                 require('components/pill/__definition__').default,
-  'pod':                  require('components/pod/__definition__').default,
+  'pod':                  require('__deprecated__/components/pod/__definition__').default,
   'portrait':             require('components/portrait/__definition__').default,
   'preview':              require('components/preview/__definition__').default,
   'profile':              require('components/profile/__definition__').default,
@@ -56,7 +56,7 @@ export default {
   'show-edit-pod':        require('components/show-edit-pod/__definition__').default,
   'sidebar':              require('components/sidebar/__definition__').default,
   'simple-color-picker':  require('components/simple-color-picker/__definition__').default,
-  'spinner':              require('components/spinner/__definition__').default,
+  'spinner':              require('__deprecated__/components/spinner/__definition__').default,
   'split-button':         require('components/split-button/__definition__').default,
   'step-sequence':        require('components/step-sequence/__definition__').default,
   'step-sequence-item':   require('components/step-sequence/step-sequence-item/__definition__').default,
