+++ src/__deprecated__/components/form/form-summary/form-summary.spec.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import TestRenderer from 'react-test-renderer';
-import { elementsTagTest, rootTagTest } from '../../../utils/helpers/tags/tags-specs/tags-specs';
-import Icon from '../../icon';
+import { elementsTagTest, rootTagTest } from '../../../../utils/helpers/tags/tags-specs/tags-specs';
+import Icon from '../../../../components/icon';
 import FormSummary from './form-summary.component';
 import 'jest-styled-components';
-import { assertStyleMatch } from '../../../__spec_helper__/test-utils';
+import { assertStyleMatch } from '../../../../__spec_helper__/test-utils';
 
 describe('<FormSummary />', () => {
   let wrapper;
