+++ src/__deprecated__/components/pod/pod.scss
@@ -1,4 +1,4 @@
-@import "./../../style-config/colors";
+@import "../../../style-config/colors.scss";
 
 .carbon-pod__block {
   background-color: $white;
