+++ src/__deprecated__/components/spinner/README.md
