+++ src/__deprecated__/components/spinner/__definition__.js
