+++ src/__deprecated__/components/spinner/classicSpinnerConfig.js
