+++ src/__deprecated__/components/spinner/docgenInfo.json
