+++ src/__deprecated__/components/spinner/documentation/index.js
