+++ src/__deprecated__/components/spinner/documentation/info.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { StoryHeader, StoryCode } from '../../../../.storybook/style/storybook-info.styles';
+import { StoryHeader, StoryCode } from '../../../../../.storybook/style/storybook-info.styles';
 
 const info = (
   <div>
