+++ src/__deprecated__/components/spinner/documentation/notes.md
