+++ src/__deprecated__/components/spinner/index.js
