+++ src/__deprecated__/components/spinner/spinner.component.js
@@ -1,11 +1,18 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import tagComponent from '../../utils/helpers/tags';
-import OptionsHelper from '../../utils/helpers/options-helper';
+import tagComponent from '../../../utils/helpers/tags';
+import OptionsHelper from '../../../utils/helpers/options-helper';
 import StyledSpinner from './spinner.style';
+import warnOfDeprecation from '../../../utils/helpers/warn-as-deprecated';
+
+let deprecatedWarnTriggered = false;
 
 const Spinner = (props) => {
   const { as, ...spinnerProps } = props;
+  if (!deprecatedWarnTriggered) {
+    deprecatedWarnTriggered = true;
+    warnOfDeprecation('Spinner', '/components/loader');
+  }
   return (
     <StyledSpinner
       type={ as } { ...spinnerProps }
