+++ src/components/filter/filter.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import Form from '../form';
+import Form from '../../__deprecated__/components/form';
 import tagComponent from '../../utils/helpers/tags';
 import './filter.scss';
