+++ src/components/show-edit-pod/show-edit-pod.js
@@ -4,8 +4,8 @@ import classNames from 'classnames';
 import I18n from 'i18n-js';
 import ReactDOM from 'react-dom';
 import CSSTransitionGroup from 'react-transition-group/CSSTransitionGroup';
-import Pod from '../pod';
-import Form from '../form';
+import Pod from '../../__deprecated__/components/pod';
+import Form from '../../__deprecated__/components/form';
 import Link from '../link';
 import Events from '../../utils/helpers/events';
 import { validProps } from '../../utils/ether';
