+++ src/utils/decorators/input/__spec__.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import TestUtils from 'react-dom/test-utils';
 import Input from './input';
-import { FormWithoutValidations as Form } from './../../../components/form';
+import { FormWithoutValidations as Form } from '../../../__deprecated__/components/form';
 import Icon from './../../../components/icon';
 import Help from './../../../components/help';
 import { shallow } from 'enzyme';
